refactor: move levels file outside states directory
